+++ templates/skeleton/app/components/Layout.tsx
@@ -1,5 +1,7 @@
 import {NavLink} from '@remix-run/react';
-import {useCustomer} from '@shopify/hydrogen';
+import {useMemo} from 'react';
+import {useMatches} from '@remix-run/react';
+import type {Customer as CustomerType} from '@shopify/hydrogen/storefront-api-types';
 
 interface LayoutProps {
   children?: React.ReactNode;
@@ -49,4 +51,20 @@ export function Layout({children, title, description}: LayoutProps) {
       {children}
     </div>
   );
+}
+
+export function useMatchesData(id: string) {
+  const matchingRoutes = useMatches();
+  const route = useMemo(
+    () => matchingRoutes.find((route) => route.id === id),
+    [matchingRoutes, id],
+  );
+
+  return route?.data;
+}
+
+export function useCustomer(): CustomerType | undefined {
+  const data = useMatchesData('root');
+
+  return data?.customer;
 }
